fix: Ajustar layout y toolbar reutilizable
