+++ src/app/app.component.html
@@ -1,15 +1,1 @@
-<mat-sidenav-container>
-
-  <mat-sidenav #sidenav>
-    <app-sidenav-list (closeSideNav)="sidenav.close()"></app-sidenav-list>
-  </mat-sidenav>
-
-  <mat-sidenav-content>
-    <app-toolbar (SideNavToggle)="sidenav.toggle()"></app-toolbar>
-
-    <main>
-      <dashboard></dashboard>
-    </main>
-
-  </mat-sidenav-content>
-</mat-sidenav-container>
+<router-outlet></router-outlet>
