+++ src/app/app.component.spec.ts
@@ -1,50 +1,38 @@
 import { TestBed, async } from '@angular/core/testing';
-import {HttpClient} from "@angular/common/http";
-import {HttpClientTestingModule} from "@angular/common/http/testing";
-
-import { AppComponent } from './app.component';
-import {HttpLoaderFactory} from "./app.module";
-import { TranslateLoader, TranslateModule, TranslateService } from '@ngx-translate/core';
-import { MaterialModule } from './modules/material.module';
 import { Router } from '@angular/router';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { MaterialModule } from 'src/app/modules/material.module';
+import { AppComponent } from './app.component';
 
-
-describe('AppComponent', () => {
-
+describe('DashboardComponent', () => {
   beforeEach(async(() => {
     TestBed.configureTestingModule({
+      imports: [
+        MaterialModule
+      ],
       declarations: [
         AppComponent
       ],
-      imports: [
-        BrowserAnimationsModule,
-        MaterialModule,
-        HttpClientTestingModule,
-        TranslateModule.forRoot({
-          loader: {
-            provide: TranslateLoader,
-            useFactory: HttpLoaderFactory,
-            deps: [HttpClient]
-          }
-        })
-      ],
       providers: [
-        TranslateService,
         {
           provide: Router,
           // useClass: class {
           //   navigate = jest.fn();
           //   navigateByUrl = jest.fn();
           // },
-        }
+        },
       ]
     }).compileComponents();
   }));
 
-  it('should create the app', async(() => {
+  it('should create the app', () => {
     const fixture = TestBed.createComponent(AppComponent);
-    const app = fixture.debugElement.componentInstance;
+    const app = fixture.componentInstance;
     expect(app).toBeTruthy();
-  }));
+  });
+
+  it(`should have as title 'atMovies'`, () => {
+    const fixture = TestBed.createComponent(AppComponent);
+    const app = fixture.componentInstance;
+    expect(app.title).toEqual('atMovies');
+  });
 });
