--- src/app/components/home/dashboard.component.html
@@ -1,2 +1,0 @@
-
-<router-outlet></router-outlet>
