--- src/app/components/home/dashboard.component.spec.ts
@@ -1,38 +1,0 @@
-import { TestBed, async } from '@angular/core/testing';
-import { Router } from '@angular/router';
-import { MaterialModule } from 'src/app/modules/material.module';
-import { DashboardComponent } from './dashboard.component';
-
-describe('DashboardComponent', () => {
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      imports: [
-        MaterialModule
-      ],
-      declarations: [
-        DashboardComponent
-      ],
-      providers: [
-        {
-          provide: Router,
-          // useClass: class {
-          //   navigate = jest.fn();
-          //   navigateByUrl = jest.fn();
-          // },
-        },
-      ]
-    }).compileComponents();
-  }));
-
-  it('should create the app', () => {
-    const fixture = TestBed.createComponent(DashboardComponent);
-    const app = fixture.componentInstance;
-    expect(app).toBeTruthy();
-  });
-
-  it(`should have as title 'atMovies'`, () => {
-    const fixture = TestBed.createComponent(DashboardComponent);
-    const app = fixture.componentInstance;
-    expect(app.title).toEqual('atMovies');
-  });
-});
