+++ src/app/components/sidenav-list/sidenav-list.component.html
@@ -1,4 +1,5 @@
 <mat-nav-list class="sidenav-list">
+  <h2>{{ 'HOME.MENU' | translate }}</h2>
   <a mat-list-item [routerLink]="['/movies/search']" (click)="onToggleClose()">
     <mat-icon>input</mat-icon>
     <span class="js-movies">{{ 'HOME.PELICULAS' | translate }}</span>
