+++ src/app/components/sidenav-list/sidenav-list.component.spec.ts
@@ -8,10 +8,13 @@ import { MatListModule } from  '@angular/material/list';
 import { MatIconModule } from '@angular/material/icon';
 import { TranslateLoader, TranslateModule, TranslateService } from '@ngx-translate/core';
 import { SidenavListComponent } from './sidenav-list.component';
-import { HttpLoaderFactory } from 'src/app/app.module';
-
+import { TranslateHttpLoader } from '@ngx-translate/http-loader';
 
 const TRANSLATIONS_ES = require('../../../assets/i18n/es.json');
+// AoT requires an exported function for factories
+export function HttpLoaderFactory(httpClient: HttpClient) {
+  return new TranslateHttpLoader(httpClient);
+}
 
 describe('SidenavListComponent', () => {
   let translate: TranslateService;
