+++ src/app/components/toolbar/toolbar.component.html
@@ -1,10 +1,14 @@
 <mat-toolbar color="primary">
   <div>
-    <button mat-icon-button (click)="openSidenav()">
-      <mat-icon>menu</mat-icon>
+    <button [hidden]="!isMain" mat-icon-button (click)="openSidenav()">
+      <mat-icon>{{isMain ? 'menu' : 'arrow_back'}}</mat-icon>
+    </button>
+
+    <button [hidden]="isMain" mat-icon-button [routerLink]="['/movies/search']">
+      <mat-icon>{{isMain ? 'menu' : 'arrow_back'}}</mat-icon>
     </button>
   </div>
 
 
-  <h1 class="app-name">{{ 'HOME.TITULO' | translate }}</h1>
+  <h1 class="app-name">{{ title }}</h1>
 </mat-toolbar>
