+++ src/app/modules/dashboard/dashboard.component.spec.ts
@@ -1,0 +1,56 @@
+import { TestBed, async } from '@angular/core/testing';
+import {HttpClient} from "@angular/common/http";
+import {HttpClientTestingModule} from "@angular/common/http/testing";
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { Router } from '@angular/router';
+
+import { DashboardComponent } from './dashboard.component';
+import { MaterialModule } from 'src/app/modules/material.module';
+
+import { TranslateLoader, TranslateModule, TranslateService } from '@ngx-translate/core';
+import { TranslateHttpLoader } from '@ngx-translate/http-loader';
+
+// AoT requires an exported function for factories
+export function HttpLoaderFactory(httpClient: HttpClient) {
+  return new TranslateHttpLoader(httpClient);
+}
+
+
+describe('AppComponent', () => {
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [
+        DashboardComponent
+      ],
+      imports: [
+        BrowserAnimationsModule,
+        MaterialModule,
+        HttpClientTestingModule,
+        TranslateModule.forRoot({
+          loader: {
+            provide: TranslateLoader,
+            useFactory: HttpLoaderFactory,
+            deps: [HttpClient]
+          }
+        })
+      ],
+      providers: [
+        TranslateService,
+        {
+          provide: Router,
+          // useClass: class {
+          //   navigate = jest.fn();
+          //   navigateByUrl = jest.fn();
+          // },
+        }
+      ]
+    }).compileComponents();
+  }));
+
+  it('should create the app', async(() => {
+    const fixture = TestBed.createComponent(DashboardComponent);
+    const app = fixture.debugElement.componentInstance;
+    expect(app).toBeTruthy();
+  }));
+});
